+++ js/data/create-offers.js
@@ -1,8 +1,9 @@
 import { SIMILAR_OFFERS } from './constants.js';
 import { createOfferData } from './create-offer-data.js';
 
-const createOffers = () => Array.from({length: SIMILAR_OFFERS }, createOfferData);
+const offers = Array.from({length: SIMILAR_OFFERS }, createOfferData);
+// const object =  Array.from({length: SIMILAR_OFFERS }, createOfferData);
 
 export {
-  createOffers
+  offers
 };
