+++ js/main.js
@@ -1,1 +1,4 @@
-import './createMapItem.js';
+import { offers } from './data/create-offers.js';
+import { createMapItems} from './create-map-items.js';
+
+createMapItems(offers);
